Refactor S3Object and PutObjectStreamWrapper
Move S3Object and PutObjectStreamWrapper as part
of s3dataset and update imports.
