+++ s3dataset/src/s3dataset/_s3client/__init__.py
@@ -1,4 +1,6 @@
 from ._s3client import S3Client
 from ._mock_s3client import MockS3Client
+from .s3object import S3Object
+from .put_object_stream_wrapper import PutObjectStreamWrapper
 
 __all__ = ["S3Client", "MockS3Client"]
