+++ s3dataset_s3_client/python/src/s3dataset_s3_client/__init__.py
@@ -3,7 +3,6 @@ import copyreg
 from ._logger_patch import TRACE as LOG_TRACE
 from ._logger_patch import _install_trace_logging
 from ._s3dataset import S3DatasetException
-from .s3object import S3Object
 
 # TODO - Find a better name than `s3dataset_s3_client`
 
@@ -17,7 +16,6 @@ def _s3dataset_exception_reduce(exc: S3DatasetException):
 copyreg.pickle(S3DatasetException, _s3dataset_exception_reduce)
 
 __all__ = [
-    "S3Object",
     "LOG_TRACE",
     "S3DatasetException",
 ]
